wildlife-monitoring-ai/requirements.txt
streamlit
opencv-python
torch
ultralytics
requests
pandas
pyserial
twilio
python-dotenv